% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods_NPAC.R
\docType{methods}
\name{getEGrowth}
\alias{getEGrowth}
\alias{getEGrowth,MizerParams_NPAC,matrix,numeric,matrix,matrix-method}
\title{getEGrowth method for the size based model}
\usage{
getEGrowth(object, n, n_pp, e_spawning, e)

\S4method{getEGrowth}{MizerParams_NPAC,matrix,numeric,matrix,matrix}(object, n,
  n_pp, e_spawning, e)
}
\arguments{
\item{object}{A \linkS4class{MizerParams_NPAC} object.}

\item{n}{A matrix of species abundance (species x size).}

\item{n_pp}{A vector of the background abundance by size.}

\item{e_spawning}{The energy available for spawning (optional, although if
specified, e must also be specified). A matrix of size no. species x no.
size bins. If not supplied, is calculated internally using the
\code{\link{getESpawning}} method.}

\item{e}{The energy available for reproduction and growth (optional, although
if specified, e_spawning must also be specified). A matrix of size no.
species x no. size bins. If not supplied, is calculated internally using
the \code{\link{getEReproAndGrowth}} method.}

\item{ocean_temp}{A numeric vector of the ocean temperature by realm or a 
single numeric ocean_temp value which is used for all realms. (PW)}
}
\value{
A two dimensional array (prey species x prey size)
}
\description{
Calculates the energy \eqn{g_i(w)} available by species and size for growth
after metabolism, movement and reproduction have been accounted for. Used by
the \code{\link{project_NPAC}} method for performing simulations.

\code{getEGrowth} method with \code{e_spawning} and \code{e} arguments.
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC, inter)
# Project with constant fishing effort for all gears for 20 time steps
sim <- project_NPAC(params, t_max = 20, effort = 0.5)
# Get the energy at a particular time step
getEGrowth(params,sim@n[21,,],sim@n_pp[21,])
}
}
\seealso{
\code{\link{project_NPAC}}
}

