% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods_NPAC.R
\docType{methods}
\name{getM2}
\alias{getM2}
\alias{getM2,MizerParams_NPAC,matrix,numeric,numeric,matrix,matrix,array-method}
\alias{getM2,MizerSim_NPAC,missing,missing,missing,missing,missing,missing-method}
\title{getM2 method for the size based model}
\usage{
getM2(object, n, n_pp, ocean_temp, phi_prey, feeding_level, pred_rate, ...)


  \S4method{getM2}{MizerParams_NPAC,matrix,numeric,numeric,matrix,matrix,array}(object,
  n, n_pp, ocean_temp, phi_prey, feeding_level, pred_rate)


  \S4method{getM2}{MizerSim_NPAC,missing,missing,missing,missing,missing,missing}(object,
  ocean_temp, time_range = dimnames(object@ocean_temp)$time, drop = TRUE,
  ...)
}
\arguments{
\item{object}{A \code{MizerParams_NPAC} or \code{MizerSim_NPAC} object.}

\item{n}{A matrix of species abundance (species x size). Only used if
\code{object} argument is of type \code{MizerParams_NPAC}.}

\item{n_pp}{A vector of the background abundance by size. Only used if
\code{object} argument is of type \code{MizerParams_NPAC}.}

\item{ocean_temp}{A numeric vector of the ocean temperature by realm or a
single numeric ocean_temp value which is used for all realms. (PW)}

\item{pred_rate}{An array of predation rates of dimension no. sp x no.
community size bins x no. of size bins in whole spectra (i.e. community +
background, the w_full slot). The array is optional. If it is not provided
it is calculated by the \code{getPredRate()} method.}

\item{...}{Other arguments (currently unused).}

\item{time_range}{Subset the returned fishing mortalities by time. The time
range is either a vector of values, a vector of min and max time, or a
single value. Default is the whole time range. Only used if the
\code{object} argument is of type \code{MizerSim_NPAC}.}

\item{drop}{Only used when object is of type \code{MizerSim_NPAC}. Should
dimensions of length 1 in the output be dropped, simplifying the output.
Defaults to TRUE}
}
\value{
If a \code{MizerParams_NPAC} object is passed in, the method returns a two
  dimensional array (prey species x prey size) based on the abundances also
  passed in. If a \code{MizerSim_NPAC} object is passed in, the method returns a
  three dimensional array (time step x prey species x prey size) with the
  predation mortality calculated at every time step in the simulation.
}
\description{
Calculates the total predation mortality \eqn{\mu_{p,i}(w_p)} on each prey
species by prey size. This method is used by the \code{\link{project_NPAC}} method
for performing simulations.

\code{getM2} method for \code{MizerParams_NPAC} object with \code{pred_rate} argument.

\code{getM2} method for \code{MizerSim_NPAC} object.
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC, inter)
# With constant fishing effort for all gears for 20 time steps
sim <- project_NPAC(params, t_max = 20, effort = 0.5)
# Get M2 at one time step
n <- sim@n[21,,]
n_pp <- sim@n_pp[21,]
getM2(params,n,n_pp)
# Get M2 at all saved time steps
getM2(sim)
# Get M2 over the time 15 - 20
getM2(sim, time_range = c(15,20))
}
}
\seealso{
\code{\link{getPredRate}} and \code{\link{project_NPAC}}.
}

