% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_NPAC.R
\docType{methods}
\name{getSSB}
\alias{getSSB}
\alias{getSSB,MizerSim_NPAC-method}
\title{Calculate the SSB of species}
\usage{
getSSB(object)

\S4method{getSSB}{MizerSim_NPAC}(object)
}
\arguments{
\item{object}{An object of class \code{MizerSim_NPAC}.}
}
\value{
An array containing the SSB (time x species)
}
\description{
Calculates the spawning stock biomass (SSB) through time of the species in
the \code{MizerSim_NPAC} class. SSB is calculated as the total mass of all mature
individuals.

Calculate the SSB from a \code{MizerSim_NPAC} object
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter_NPAC)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC, inter_NPAC)
# With constant fishing effort for all gears for 20 time steps
sim <- project_NPAC(params, t_max = 20, effort = 0.5)
getSSB(sim)
}
}

