% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods_NPAC.R
\docType{methods}
\name{getTempEffect}
\alias{getTempEffect}
\alias{getTempEffect,MizerParams_NPAC,matrix-method}
\alias{getTempEffect,MizerParams_NPAC,numeric-method}
\alias{getTempEffect,MizerSim_NPAC,missing-method}
\title{Get the total temperature effect across all realms by time, species, and size.}
\usage{
getTempEffect(object, ocean_temp, ...)

\S4method{getTempEffect}{MizerParams_NPAC,numeric}(object, ocean_temp, ...)

\S4method{getTempEffect}{MizerParams_NPAC,matrix}(object, ocean_temp, ...)

\S4method{getTempEffect}{MizerSim_NPAC,missing}(object, ocean_temp,
  time_range = dimnames(object@ocean_temp)$time, drop = TRUE, ...)
}
\arguments{
\item{object}{A \code{MizerParams_NPAC} object or a \code{MizerSim_NPAC} object}

\item{ocean_temp}{The ocean temperature in each realm. Only needed if the object
argument is of class \code{MizerParams_NPAC}. See notes below.}

\item{...}{Other arguments passed to \code{getTempEffect} method.}

\item{time_range}{Subset the returned temperature effects by time. The time
range is either a vector of values, a vector of min and max time, or a
single value. Default is the whole time range. Only used if the
\code{object} argument is of type \code{MizerSim_NPAC}.}

\item{drop}{Only used when object is of type \code{MizerSim_NPAC}. Should
dimensions of length 1 be dropped, e.g., if your community has only one
species it might make presentation of results easier. Default is TRUE.}
}
\value{
An array. If the effort argument has a time dimension, or object is
  of class \code{MizerSim_PW}, the output array has three dimensions (time x 
  species x size). If the ocean_temp argument does not have a time dimension, the
  output array has two dimensions (species x size).
}
\description{
Calculates the temperature effect across all realms by species and size at 
each time step in teh \code{ocean_temp} argument.
The total temperature effect is just the sum of the temperature effects in
each realm (PW)

\code{getTempEffect} method for \code{MizerParams_PW} object with constant temperature.

\code{getTempEffect} method for \code{MizerParams_NPAC} object with time changing effort.

\code{getTempEffect} method for \code{MizerSim_NPAC} object.
}
\note{
Here: temperature effect is determined from exposure x ontogenetic_migration
  x ocean_temp. (PW)

The \code{ocean_temp} argument is only used if a \code{MizerParams_NPAC} object is
  passed in. The \code{ocean_temp} argument can be a two dimension array (time x 
  realm), a vector of length equal to the number of realms (each realm has a 
  different temperature that is constant in time), or a single numeric value 
  (each realm has the same temperature that is constant in time). The order of 
  realms in the \code{ocean_temp} argument must be the same as in the 
  \code{MizerParams_NPAC} object. (PW)

If the object argument is of class \code{MizerSim_NPAC} then the ocean_temp slot
  of the \code{MizerSim_NPAC} object is used and the \code{ocean_temp} argument 
  is not used. (PW)
}

