% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods_NPAC.R
\docType{methods}
\name{getTempEffectRealm}
\alias{getTempEffectRealm}
\alias{getTempEffectRealm,MizerParams_NPAC,matrix-method}
\alias{getTempEffectRealm,MizerParams_NPAC,numeric-method}
\alias{getTempEffectRealm,MizerSim_NPAC,missing-method}
\title{getTempEffectRealm method for the size based model (PW)}
\usage{
getTempEffectRealm(object, ocean_temp, ...)

\S4method{getTempEffectRealm}{MizerParams_NPAC,numeric}(object, ocean_temp, ...)

\S4method{getTempEffectRealm}{MizerParams_NPAC,matrix}(object, ocean_temp, ...)

\S4method{getTempEffectRealm}{MizerSim_NPAC,missing}(object, ocean_temp,
  time_range = dimnames(object@ocean_temp)$time, ...)
}
\arguments{
\item{object}{A \code{MizerParams_NPAC} object or a \code{MizerSim_NPAC}
object.}

\item{ocean_temp}{The ocean temperature in each realm. Only needed if the object 
argument is of class \code{MizerParams_NPAC}. See notes below.}

\item{...}{Other arguments (currently unused)/ (PW)}

\item{time_range}{Subset the returned temperature effects by time. The time
range is either a vector of values, a vector of min and max time, or a 
single value. Default is the whole time range. Only used if the \code{object}
argument is of type \code{MizerSim_NPAC}.}
}
\value{
An array. If the ocean_temp argument has a time dimension, or a 
  \code{MizerSim_NPAC} is passed in, the output array has four dimensions
  (time x realm x species x size). If the ocean_temp argument does not have a 
  time dimension (i.e., it is a vector or a single numeric), the output 
  array has three dimensions (realm x species x size).
}
\description{
Calcuates the temperature effect by realm, species, and size at each time step
in the \code{ocean_temp} argument. Used by the \code{project_NPAC} method to 
perform simulations (PW)

\code{getTempEffectRealm} method for \code{MizerParams_NPAC} object with constant temperature. (PW)

\code{getTempEffectRealm} method for \code{MizerParams_NPAC} object with time-changing temp.

\code{getTempEffectRealm} method for \code{MizerSim_PW} object.
}
\note{
Here: Temperature effect results from exposure x ontogenetic migration
  x ocean temperature. (PW)

The \code{ocean_temp} argument is only used of a \code{MizerParams_NPAC} 
  object is passed in. The \code{ocean_temp} argument can be a two dimensional
  array (time x realm), a vector of length equal to the number of realms 
  (each realm has a different temperature that is constant in time), or a
  single numeric value (each realm has the same temperature that is constant
  in time). The order of realms in the \code{ocean_temp} argument must be the
  same as in the \code{MizerParams_NPAC} object. (PW)

If the object argument is of class \code{MizerSim_NPAC} then the 
  ocean_temp slot of the \code{MizerSim_NPAC} object is used and the 
  \code{ocean_temp} argument is not used.
}

