% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_NPAC.R
\docType{methods}
\name{getYield}
\alias{getYield}
\alias{getYield,MizerSim_NPAC-method}
\title{Calculate the total yield of each species}
\usage{
getYield(object)

\S4method{getYield}{MizerSim_NPAC}(object)
}
\arguments{
\item{object}{An object of class \code{MizerSim_NPAC}.}
}
\value{
An array containing the total yield (time x species)
}
\description{
Calculates the total yield of each species across all gears at each
simulation time step.

Calculate the total yield of each species from a \code{MizerSim_NPAC} object
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter_NPAC)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC, inter_NPAC)
sim <- project_NPAC(params, effort=1, t_max=10)
y <- getYield(sim)
}
}
\seealso{
\code{\link{getYieldGear}}
}

