% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_NPAC.R
\docType{package}
\name{mizerNPAC}
\alias{mizerNPAC}
\alias{mizerNPAC,}
\alias{mizerNPAC-package}
\title{mizerNPAC: Multi-species size-based modelling in R}
\description{
The mizerNPAC package implements multi-species size-based modelling in R. It has 
been designed for modelling marine ecosystems.
}
\details{
Using \pkg{mizerNPAC} is relatively simple. There are three main stages: 
\enumerate{

\item Setting the model parameters. This is done by creating an object of 
class \code{MizerParams_NPAC}. This includes model parameters such as the life 
history parameters of each species, and the range of the size spectrum.

\item Running a simulation. This is done by calling the \code{project_NPAC()} 
method on the model parameters. This produces an object of \code{MizerSim_NPAC} 
which contains the results of the simulation.

\item Exploring results. After a simulation has been run, the results can be 
explored using a range of plots and summaries. 
}

See the accompanying vignette for full details of the principles behind mizerNPAC
and how the package can be used to peform size-based modelling.
}

