% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim_NPAC-class.R
\docType{class}
\name{MizerSim_NPAC-class}
\alias{MizerSim_NPAC-class}
\title{MizerSim_NPAC}
\description{
A class that holds the results of projecting a \linkS4class{MizerParams_NPAC}
object through time.
}
\details{
\code{MizerSim_NPAC} objects are created by using the \code{\link{project_NPAC}} method
on an object of type \code{MizerParams_NPAC}.

There are several plotting methods available to explore the contents of a
\code{MizerSim_NPAC} object. See the package vignette for more details.
}
\section{Slots}{

\describe{
\item{\code{params}}{An object of type \linkS4class{MizerParams_NPAC}.}

\item{\code{n}}{Array that stores the projected community population abundances by
time, species and size}

\item{\code{effort}}{Array that stores the fishing effort through time by time and
gear}

\item{\code{ocean_temp}}{Array that stores the ocean temperature through time by
time and realm (PW)}

\item{\code{n_pp}}{Array that stores the projected background population by time and
size}
}}
\seealso{
\code{\link{project_NPAC}} \code{\link{MizerParams_NPAC}}
}

