% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MizerSim_NPAC-class.R
\docType{methods}
\name{MizerSim_NPAC}
\alias{MizerSim_NPAC}
\alias{MizerSim_NPAC,MizerParams_NPAC-method}
\title{Constructor for the \code{MizerSim_NPAC} class}
\usage{
MizerSim_NPAC(object, ...)

\S4method{MizerSim_NPAC}{MizerParams_NPAC}(object, t_dimnames = NA,
  t_max = 100, t_save = 1)
}
\arguments{
\item{object}{a \linkS4class{MizerParams_NPAC} object}

\item{...}{Other arguments (currently not used).}

\item{t_dimnames}{Numeric vector that is used for the time dimensions of the
slots. Default = NA.}

\item{t_max}{The maximum time step of the simulation. Only used if t_dimnames
= NA. Default value = 100.}

\item{t_save}{How often should the results of the simulation be stored. Only
used if t_dimnames = NA. Default value = 1.}
}
\value{
An object of type \linkS4class{MizerSim_NPAC}
}
\description{
A constructor for the \code{MizerSim_NPAC} class. This is used by the
\code{project_NPAC} method to create \code{MizerSim_NPAC} objects of the right
dimensions. It is not necessary for users to use this constructor.

MizerSim_NPAC constructor taking only a \code{MizerParams_NPAC} object.
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter_NPAC)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC, inter_NPAC)
sim <- project_NPAC(params)
}
}
\seealso{
\code{\link{project_NPAC}} \linkS4class{MizerParams_NPAC}
  \linkS4class{MizerSim_NPAC}
}

