% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_NPAC.R
\docType{methods}
\name{plotTempEffect}
\alias{plotTempEffect}
\alias{plotTempEffect,MizerSim_NPAC-method}
\title{Plot total temperature effect on each species by size (PW)}
\usage{
plotTempEffect(object, ...)

\S4method{plotTempEffect}{MizerSim_NPAC}(object,
  time_range = max(as.numeric(dimnames(object@n)$time)), print_it = TRUE,
  ...)
}
\arguments{
\item{object}{An object of class \code{MizerSim_NPAC}.}

\item{...}{Other arguments to pass to \code{getFMort} method}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{print_it}{Display the plot, or just return the ggplot2 object}
}
\value{
A ggplot2 object
}
\description{
After running a projection, plot the total temperature effect on each species
by size. The total temperature effect is averaged over the specified time
range (a single value for the time range can be used to plot a single time
step).

Plot total temperature effect using a \code{MizerSim_NPAC} object. (PW)
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter_NPAC)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC, inter_NPAC)
sim <- project_NPAC(params, effort=1, t_max=20, t_save = 2)
plotTempEffect(sim)
plotTempEffect(sim, time_range = 10:20)
}
}
\seealso{
\code{\link{getTempEffect}}
}

