% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_NPAC.R
\docType{methods}
\name{summary,MizerParams_NPAC-method}
\alias{summary,MizerParams_NPAC-method}
\title{Summarize MizerParams_NPAC object}
\usage{
\S4method{summary}{MizerParams_NPAC}(object, ...)
}
\arguments{
\item{object}{A \code{MizerParams_NPAC} object.}

\item{...}{Other arguments (currently not used).}
}
\description{
Outputs a general summary of the structure and content of the object
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter_NPAC)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC,inter_NPAC)
summary(params)
}
}

