% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods_NPAC.R
\docType{methods}
\name{summary,MizerSim_NPAC-method}
\alias{summary,MizerSim_NPAC-method}
\title{Summarize MizerSim_NPAC object}
\usage{
\S4method{summary}{MizerSim_NPAC}(object, ...)
}
\arguments{
\item{object}{A \code{MizerSim_NPAC} object.}

\item{...}{Other arguments (currently not used).}
}
\description{
Outputs a general summary of the structure and content of the object
}
\examples{
\dontrun{
data(NS_species_params_gears_NPAC)
data(inter_NPAC)
params <- MizerParams_NPAC(NS_species_params_gears_NPAC,inter_NPAC)
sim <- project_NPAC(params, effort=1, t_max=5)
summary(sim)
}
}

